requirements.txt
dj-database-url==0.4.1
Django==1.11.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
django-flat-responsive==1.4.1
requests==2.18.4
rq==0.8.2
djangorestframework
